# backend/server/swagger.yaml
openapi: 3.0.3
info:
  title: Pixel Protocol API
  description: |
    Backend API for Pixel Protocol game with hero management and battles.
    
    ## Authentication
    Most endpoints require player authentication via:
    1. **Request Body**: Include `_secret` field with player's UUID secret (primary method)
    2. **Cookie**: Use `player_secret` cookie set by hero creation (fallback method)
    
    When creating a hero, a secure `player_secret` cookie is automatically set for subsequent requests.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.pixelprotocol.com
    description: Production server
paths:
  /api/player:
    post:
      summary: Create a new player
      tags:
        - Player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  description: Player's username
                  example: "JohnDoe"
      responses:
        '201':
          description: Player created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Invalid request
        '409':
          description: Username already exists with this suffix
        '500':
          description: Internal server error
  /api/player/{id}:
    get:
      summary: Get player by ID
      tags:
        - Player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Player UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretRequest'
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Player not found
          
    put:
      summary: Update player (full update)
      tags:
        - Player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - _secret
              properties:
                username:
                  type: string
                _secret:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Player updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Player not found
        '409':
          description: Username conflict
          
    patch:
      summary: Update player (partial update)
      tags:
        - Player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - _secret
              properties:
                username:
                  type: string
                _secret:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Player updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Player not found
          
    delete:
      summary: Soft delete player
      tags:
        - Player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretRequest'
      responses:
        '204':
          description: Player deleted
        '404':
          description: Player not found
...
components:
  securitySchemes:
    PlayerSecret:
      type: apiKey
      in: header
      name: Cookie
      description: |
        Player authentication via secure cookie. Cookie is automatically set when creating a hero.
        Cookie name: `player_secret`
    PlayerSecretBody:
      type: http
      scheme: bearer
      description: |
        Player authentication via request body. Include `_secret` field with player's UUID secret.
        Takes precedence over cookie authentication.
  schemas:
    Player:
      type: object
      properties:
        ID:
          type: string
          format: uuid
        UserName:
          type: string
        UserNameSuffix:
          type: integer
          format: int32
        Secret:
          type: string
          format: uuid
        DeletedAt:
          type: string
          format: date-time
          nullable: true
...
    SecretRequest:
      type: object
      required:
        - _secret
      properties:
        _secret:
          type: string
          format: uuid
          description: Player's secret for authentication
...